ltg/requirements.txt
defusedxml==0.4.1
Django==1.9.5
djangorestframework==3.3.3
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-social-auth==0.2.18
python3-openid==3.0.10
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
